% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CRDC_gene_info_page_url.R
\name{get_CRDC_gene_info_page_url}
\alias{get_CRDC_gene_info_page_url}
\title{Get Gene Info Page URL in CRDC}
\usage{
get_CRDC_gene_info_page_url(gene_id)
}
\arguments{
\item{gene_id}{a gene ID string, e.g. "Os05g0187500".}
}
\value{
corresponding gene info page URL in CRDC.
}
\description{
Get Gene Info Page URL in CRDC
}
