% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_realplex.R
\name{read_realplex}
\alias{read_realplex}
\title{Reading realplex PCR result .XLS file}
\usage{
read_realplex(xlsfile, wellfile)
}
\arguments{
\item{xlsfile}{realplex result .XLS file.}
}
\value{
a data.frame, resembling 96-well plate.
}
\description{
Reading realplex PCR result .XLS file
}
