% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wseen.R
\name{read_wseen}
\alias{read_wseen}
\title{Read WSeen Seed Scan Data}
\usage{
read_wseen(folder)
}
\arguments{
\item{folder}{WSeen output folder.}
}
\value{
a tibble containing wseen data.
}
\description{
Read WSeen Seed Scan Data
}
